Fix review findings: strongbox dupe, handless breaking, stacked cells
- Strongbox: breakBlock now empties the tile. The contents are already
  in the dropped item on every path, and a player with the box still
  open could keep taking items out of the removed tile for the rest of
  the tick - a duplicate of each also in the dropped box.
- Every block that carries state in its item (all machines, the Cell,
  tank, strongbox) now harvests with any tool or a bare hand. Iron
  material demanded a pickaxe; without one getDrops never ran and the
  augments, RF, fluid or a box's whole inventory were deleted.
- Transposer: fluid containers are split off one at a time into a
  per-line process slot (saved in NBT, spilled on break), like TE's own
  slot 1. A stack of cells used to jam a line forever. A container that
  refuses the tank's fluid now goes to the output instead of blocking.
- Tank light level computed in long: capacity * 3 overflowed past
  ~715M mB (the config allows a billion).

+++ src/main/java/net/thermaladd/mod/tileentity/TileSingularityTank.java
@@ -75,15 +75,16 @@ public class TileSingularityTank extends TileEntity implements IFluidHandler {
             return 0;
         }
         int glow = fluid.getFluid().getLuminosity(fluid);
-        int capacity = tank.getCapacity();
-        int amount = fluid.amount;
+        // In long: the configurable capacity goes up to a billion, where capacity * 3 overflows.
+        long capacity = tank.getCapacity();
+        long amount = fluid.amount;
         if (amount <= capacity / 4) {
             return glow >> 1;
         }
         if (amount >= capacity * 3 / 4) {
             return glow;
         }
-        return (glow >> 1) + (glow - (glow >> 1)) * (amount - capacity / 4) / (capacity / 2);
+        return (int) ((glow >> 1) + (glow - (glow >> 1)) * (amount - capacity / 4) / (capacity / 2));
     }
 
     // ---------------------------------------------------------------- dismantling
